requirements.txt
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
jinja2==3.1.3 